+++ native/GhostRigger.Windows.Editor.Level/Python/src/gui/windows/module_editor_window.py
@@ -2875,6 +2875,7 @@ class ModuleEditorWindow(QtWidgets.QMainWindow):
         readiness_result = self.controller.authored_module_readiness()
         self.workflow_panel.set_state(self.project, readiness_result.readiness)
         self.readiness_panel.set_readiness(readiness_result.readiness)
+        self.export_panel.set_readiness(readiness_result.readiness)
         self.validation_panel.set_issues(self.controller.validate())
         if self.controller.model.selected_ids:
             self.select_item(self.controller.model.selected_ids[0])